pyproject.toml
[tool.poetry]
name = "kivy-sailcommander"
version = "0.1.0"
description = ""
authors = ["Heiko Kulinna <Heikokulinna@gmx.de>"]
readme = "README.md"
# packages = [{include = "kivy_sailcommander"}]

[tool.poetry.dependencies]
python = ">=3.8.1, <3.9"
Kivy = ">=2.0.0"   # 2.1.0
kivymd = ">=1.1.1"   # 1.1.1
# pandas = ">=1.5.2"
# numpy = ">=1.24.1"
# kivy==2.1.0, kivymd==1.1.0, sdl2_ttf == 2.0.15, pillow


#python-dotenv = "*"   #"0.21.1"
# "yaml-1.3" = ">= 0.1.0"
pyyaml = "*" # ">= 6.0"
buildozer = "^1.5.0"
cython = "^0.29.33"


[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
flake8 = "^6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
